Update the unit test for c
